+++ app/src/main/java/dk/sdu/weshare/api/Api.kt
@@ -3,43 +3,130 @@ package dk.sdu.weshare.api
 import dk.sdu.weshare.models.Expense
 import dk.sdu.weshare.models.Group
 import dk.sdu.weshare.models.User
+import dk.sdu.weshare.util.ServiceBuilder
 import retrofit2.Call
-import retrofit2.http.Body
-import retrofit2.http.DELETE
-import retrofit2.http.GET
-import retrofit2.http.POST
-import retrofit2.http.PUT
-import retrofit2.http.Path
-import retrofit2.http.Query
+import retrofit2.Callback
+import retrofit2.Response
 
-interface Api {
-	@POST("api/groups/{id}/expenses")
-	fun addExpenseToGroup(@Path("id") groupId: Int, @Body expense: Expense): Call<Group>
+class Api {
+	companion object {
+		private val retrofit = ServiceBuilder.buildService(ApiEndpoints::class.java)
 
-	@POST("api/groups/{id}/users")
-	fun addUserToGroup(@Path("id") groupId: Int, @Query("email") userEmail: String): Call<Group>
+		private fun <T> simpleCallback(onResult: (T?) -> Unit): Callback<T> {
+			return object : Callback<T> {
+				override fun onFailure(call: Call<T>, t: Throwable) {
+					onResult(null)
+				}
 
-	@DELETE("api/groups/{id}/users")
-	fun removeUserFromGroup(@Path("id") groupId: Int, @Query("email") userEmail: String): Call<Group>
+				override fun onResponse(call: Call<T>, response: Response<T>) {
+					onResult(response.body())
+				}
+			}
+		}
 
-	@GET("api/groups/{id}")
-	fun getGroup(@Path("id") groupId: Int): Call<Group>
+		fun addExpenseToGroup(groupId: Int, expense: Expense, onResult: (Group?) -> Unit) {
+			retrofit
+				.addExpenseToGroup(groupId, expense)
+				.enqueue(simpleCallback(onResult))
+		}
 
-	@PUT("api/groups/{id}")
-	fun updateGroup(@Path("id") groupId: Int, @Query("name") name: String): Call<Group>
+		fun addUserToGroup(group: Group, user: User, onResult: (Group?) -> Unit) {
+			retrofit
+				.addUserToGroup(group.id, user.email)
+				.enqueue(simpleCallback(onResult))
+		}
+		fun addUserToGroup(group: Group, userEmail: String, onResult: (Group?) -> Unit) {
+			retrofit
+				.addUserToGroup(group.id, userEmail)
+				.enqueue(simpleCallback(onResult))
+		}
+		fun addUserToGroup(groupId: Int, user: User, onResult: (Group?) -> Unit) {
+			retrofit
+				.addUserToGroup(groupId, user.email)
+				.enqueue(simpleCallback(onResult))
+		}
+		fun addUserToGroup(groupId: Int, userEmail: String, onResult: (Group?) -> Unit) {
+			retrofit
+				.addUserToGroup(groupId, userEmail)
+				.enqueue(simpleCallback(onResult))
+		}
 
-	@DELETE("api/groups/{id}")
-	fun deleteGroup(@Path("id") groupId: Int): Call<Error>
+		fun removeUserFromGroup(group: Group, user: User, onResult: (Group?) -> Unit) {
+			retrofit
+				.removeUserFromGroup(group.id, user.email)
+				.enqueue(simpleCallback(onResult))
+		}
+		fun removeUserFromGroup(group: Group, userEmail: String, onResult: (Group?) -> Unit) {
+			retrofit
+				.removeUserFromGroup(group.id, userEmail)
+				.enqueue(simpleCallback(onResult))
+		}
+		fun removeUserFromGroup(groupId: Int, user: User, onResult: (Group?) -> Unit) {
+			retrofit
+				.removeUserFromGroup(groupId, user.email)
+				.enqueue(simpleCallback(onResult))
+		}
+		fun removeUserFromGroup(groupId: Int, userEmail: String, onResult: (Group?) -> Unit) {
+			retrofit
+				.removeUserFromGroup(groupId, userEmail)
+				.enqueue(simpleCallback(onResult))
+		}
 
-	@GET("api/groups")
-	fun getAllGroups(): Call<List<Group>>
+		fun getGroup(group: Group, onResult: (Group?) -> Unit) {
+			retrofit
+				.getGroup(group.id)
+				.enqueue(simpleCallback(onResult))
+		}
+		fun getGroup(groupId: Int, onResult: (Group?) -> Unit) {
+			retrofit
+				.getGroup(groupId)
+				.enqueue(simpleCallback(onResult))
+		}
 
-	@POST("api/groups")
-	fun createGroup(@Query("name") name: String): Call<Group>
+		fun updateGroup(group: Group, name: String, onResult: (Group?) -> Unit) {
+			retrofit
+				.updateGroup(group.id, name)
+				.enqueue(simpleCallback(onResult))
+		}
+		fun updateGroup(groupId: Int, name: String, onResult: (Group?) -> Unit) {
+			retrofit
+				.updateGroup(groupId, name)
+				.enqueue(simpleCallback(onResult))
+		}
 
-	@POST("/api/register")
-	fun register(@Query("name") name: String, @Query("email") email: String, @Query("password") password: String, @Query("device_name") deviceName: String): Call<User>
+		fun deleteGroup(group: Group, onResult: (Error?) -> Unit) {
+			retrofit
+				.deleteGroup(group.id)
+				.enqueue(simpleCallback(onResult))
+		}
+		fun deleteGroup(groupId: Int, onResult: (Error?) -> Unit) {
+			retrofit
+				.deleteGroup(groupId)
+				.enqueue(simpleCallback(onResult))
+		}
 
-	@POST("/api/login")
-	fun login(@Query("email") email: String, @Query("password") password: String, @Query("device_name") deviceName: String): Call<User>
+		fun getAllGroups(onResult: (List<Group>?) -> Unit) {
+			retrofit
+				.getAllGroups()
+				.enqueue(simpleCallback(onResult))
+		}
+
+		fun createGroup(name: String, onResult: (Group?) -> Unit) {
+			retrofit
+				.createGroup(name)
+				.enqueue(simpleCallback(onResult))
+		}
+
+		fun register(name: String, email: String, password: String, onResult: (User?) -> Unit) {
+			retrofit
+				.register(name, email, password, android.os.Build.MODEL)
+				.enqueue(simpleCallback(onResult))
+		}
+
+		fun login(email: String, password: String, onResult: (User?) -> Unit) {
+			retrofit
+				.login(email, password, android.os.Build.MODEL)
+				.enqueue(simpleCallback(onResult))
+		}
+	}
 }
